+++ spec/factories/factories.rb
@@ -1,4 +1,0 @@
-FactoryBot.define do
-  factory :paxful_engine_trade, class: "PaxfulEngine::Trade" do
-  end
-end