+++ spec/support/factory_bot.rb
@@ -1,3 +1,6 @@
+require "factory_bot_rails"
+require "paxful_engine/factories"
+
 RSpec.configure do |config|
   config.include FactoryBot::Syntax::Methods